Extracted ElasticSearchConfigurationTest and ElasticSearchTestCase
